Hard Two Pointers Problem
First of the random problems I solved. It was a daily challenge. Nearly solved it, but was obsessed with storing previously seen values to ensure we stayed at or under 100 get calls. Ultimately, this wasn't really needed. This solution isn't that efficient from a space or memory perspective (probably because we could find the answer sooner without running through all these steps), but it is easy to understand. Didn't seem that tricky to implement, although it was a hard problem.